# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Extractor",
    "description": "Brazilian economic indicators extraction, storage, and dashboard API. 25 macro indicators from BCB SGS, IBGE SIDRA, and B3/Yahoo Finance. Supports on-demand transforms, pinned dashboards, and release calendars.",
    "version": "0.5.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/series/{code}/transform": {
      "post": {
        "tags": [
          "transform"
        ],
        "summary": "Apply a transform to a series",
        "description": "Applies the requested statistical transform to all stored observations for the series. Results are cached in Redis (TTL: daily=1h, monthly=24h, quarterly=7d). Cache key includes the transform spec hash and the latest observed_at timestamp, so new observations automatically invalidate the cache. NaN positions produced by the transform are reported in `metadata.gaps`. Decimal DB values are returned as float64.",
        "operationId": "apply_transform_series__code__transform_post",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransformRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transform applied and result returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransformResponse"
                }
              }
            }
          },
          "404": {
            "description": "Series code not found."
          },
          "422": {
            "description": "Invalid TransformSpec: unknown op or invalid params. See spec for valid op values."
          },
          "503": {
            "description": "Transform computation failed (e.g. numeric overflow)."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GapRecord": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date",
            "description": "ISO-8601 date string (YYYY-MM-DD) of the gap."
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Reason for the gap. Currently always 'missing_upstream'.",
            "default": "missing_upstream"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "GapRecord",
        "description": "A single NaN gap detected in a transformed series."
      },
...
      "TransformMetadata": {
        "properties": {
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/GapRecord"
            },
            "type": "array",
            "title": "Gaps",
            "description": "List of NaN positions detected after transform. Empty when no gaps."
          },
          "stub": {
            "type": "boolean",
            "title": "Stub",
            "description": "True when the op is a stub (sa, calendar_adj) — values are level passthrough.",
            "default": false
          },
          "op": {
            "type": "string",
            "title": "Op",
            "description": "Transform operation applied."
          },
          "params": {
            "type": "object",
            "title": "Params",
            "description": "Effective parameters used."
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "True if the result was served from Redis cache."
          },
          "latest_observed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Observed At",
            "description": "ISO-8601 timestamp of the most recent observation used for cache keying."
          }
        },
        "type": "object",
        "required": [
          "op",
          "cached"
        ],
        "title": "TransformMetadata",
        "description": "Metadata about a transform computation result."
      },
      "TransformRequest": {
        "properties": {
          "op": {
            "type": "string",
            "title": "Op",
            "description": "Transform operation identifier. One of: level, sa, calendar_adj, mom, qoq, yoy, annualized, diff, log_diff, pp, ma, ewma, accum12, stddev12, rebase, zscore, percentile."
          },
          "params": {
            "type": "object",
            "title": "Params",
            "description": "Operation parameters. op-specific keys: ma → {window: int (default 12)}, ewma → {span: int (default 12)}, rebase → {base: float (default 100.0)}. All other ops ignore params."
          }
        },
        "type": "object",
        "required": [
          "op"
        ],
        "title": "TransformRequest",
        "description": "Request body for POST /series/{code}/transform."
      },
      "TransformResponse": {
        "properties": {
          "series_code": {
            "type": "string",
            "title": "Series Code",
            "description": "Series code the transform was applied to."
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/TransformValuePoint"
            },
            "type": "array",
            "title": "Values",
            "description": "Transformed time series ordered by date ascending."
          },
          "metadata": {
            "$ref": "#/components/schemas/TransformMetadata",
            "description": "Computation metadata including gap list."
          }
        },
        "type": "object",
        "required": [
          "series_code",
          "values",
          "metadata"
        ],
        "title": "TransformResponse",
        "description": "Response schema for POST /series/{code}/transform."
      },
      "TransformValuePoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date",
            "description": "ISO-8601 date string (YYYY-MM-DD)."
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "Transformed float value. null where the transform produced NaN."
          }
        },
        "type": "object",
        "required": [
          "date",
          "value"
        ],
        "title": "TransformValuePoint",
        "description": "A single (date, value) point in a transform result."
      },
...
    }
  },
...
}